server-endpoint: serve blobs by hash
Introduce a new binary that can act as an endpoint to serve blobs in a
packfile given their hashes.  Reachability checks are performed before
the packfile is sent - both an absent blob and an unreachable blob are
reported to the user in the same way ("not our blob").

Due to a bug in "rev-list" in the absence of bitmaps (discussed here
[1]), the server repositories in tests all have bitmaps.

[1] <20170309003547.6930-1-jonathantanmy@google.com>

Signed-off-by: Jonathan Tan <jonathantanmy@google.com>